+++ CHANGES.md
@@ -1,3 +1,6 @@
+#### Version 16
+- [#25](https://github.com/kotcrab/ghidra-allegrex/issues/25) - added support for type A relocations only present in program headers
+
 #### Version 15
 - Fixed debugger breakpoint placing